pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["atlast_sc*"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.yaml"]

[project]
#name = "sensitivity_calculator"
name = "atlast_sc"
version = "0.1.0"
authors = [
    { name="Joanna Ramasawmy", email="joanna.ramasawmy@stfc.ac.uk" }
]
description = "AtLast sensitivity calculator"
#license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "astropy == 5.1.*",
    "numpy == 1.23.*",
    "scipy == 1.9.*",
    "pyyaml == 6.0.*",
    "pydantic == 1.10.*"
]

[project.urls]
repository="https://github.com/ukatc/AtLAST_sensitivity_calculator"
